# Conflicts:
#	.idea/modules.xml

+++ finmartserviceapi/src/main/java/magicfinmart/datacomp/com/finmartserviceapi/finmart/model/HealthQuoteEntity.java
@@ -1,247 +1,247 @@
 package magicfinmart.datacomp.com.finmartserviceapi.finmart.model;
 
 public class HealthQuoteEntity {
-            /**
-             * CustomerReferenceID : 170067
-             * QuoteId : 22
-             * PolicyTermYear : 1
-             * InsurerName : Aditya Birla Health Insurance Company Limited.
-             * InsurerLogoName : AdityaBirla.png
-             * InsurerId : 42
-             * PlanName : Diamond
-             * PlanID : 292
-             * OtherPlanID : 20180209421970005001
-             * NetPremium : 7053
-             * ServiceTax : 836.78
-             * SumInsured : 300000
-             * HMBValue : 
-             * IsOnlinePayment : true
-             * KeyFeatures : 
-             * BroucherDownloadLink : /
-             * Discount : 0
-             * GrossPremium : 5977
-             * DiscountPercent : 
-             * Premium : 
-             * Group_name : 
-             * Deductible_Amount : 0
-             * QuoteStatus : Success
-             * ProposerPageUrl : http://qa.policyboss.com/HealthInsuranceIndia/ContactDetails?CustomerReferenceNumber=170067&SelectedQuoteId=22&SupportsAgentID=2&CallingSource=POSPAPP&IsCustomer=0
-             */
+    /**
+     * CustomerReferenceID : 170067
+     * QuoteId : 22
+     * PolicyTermYear : 1
+     * InsurerName : Aditya Birla Health Insurance Company Limited.
+     * InsurerLogoName : AdityaBirla.png
+     * InsurerId : 42
+     * PlanName : Diamond
+     * PlanID : 292
+     * OtherPlanID : 20180209421970005001
+     * NetPremium : 7053
+     * ServiceTax : 836.78
+     * SumInsured : 300000
+     * HMBValue :
+     * IsOnlinePayment : true
+     * KeyFeatures :
+     * BroucherDownloadLink : /
+     * Discount : 0
+     * GrossPremium : 5977
+     * DiscountPercent :
+     * Premium :
+     * Group_name :
+     * Deductible_Amount : 0
+     * QuoteStatus : Success
+     * ProposerPageUrl : http://qa.policyboss.com/HealthInsuranceIndia/ContactDetails?CustomerReferenceNumber=170067&SelectedQuoteId=22&SupportsAgentID=2&CallingSource=POSPAPP&IsCustomer=0
+     */
 
-            private int CustomerReferenceID;
-            private int QuoteId;
-            private int PolicyTermYear;
-            private String InsurerName;
-            private String InsurerLogoName;
-            private int InsurerId;
-            private String PlanName;
-            private int PlanID;
-            private String OtherPlanID;
-            private int NetPremium;
-            private double ServiceTax;
-            private int SumInsured;
-            private String HMBValue;
-            private boolean IsOnlinePayment;
-            private String KeyFeatures;
-            private String BroucherDownloadLink;
-            private int Discount;
-            private int GrossPremium;
-            private String DiscountPercent;
-            private String Premium;
-            private String Group_name;
-            private int Deductible_Amount;
-            private String QuoteStatus;
-            private String ProposerPageUrl;
+    private int CustomerReferenceID;
+    private int QuoteId;
+    private int PolicyTermYear;
+    private String InsurerName;
+    private String InsurerLogoName;
+    private int InsurerId;
+    private String PlanName;
+    private int PlanID;
+    private String OtherPlanID;
+    private double NetPremium;
+    private double ServiceTax;
+    private int SumInsured;
+    private String HMBValue;
+    private boolean IsOnlinePayment;
+    private String KeyFeatures;
+    private String BroucherDownloadLink;
+    private int Discount;
+    private int GrossPremium;
+    private String DiscountPercent;
+    private String Premium;
+    private String Group_name;
+    private int Deductible_Amount;
+    private String QuoteStatus;
+    private String ProposerPageUrl;
 
-            public int getCustomerReferenceID() {
-                return CustomerReferenceID;
-            }
+    public int getCustomerReferenceID() {
+        return CustomerReferenceID;
+    }
 
-            public void setCustomerReferenceID(int CustomerReferenceID) {
-                this.CustomerReferenceID = CustomerReferenceID;
-            }
+    public void setCustomerReferenceID(int CustomerReferenceID) {
+        this.CustomerReferenceID = CustomerReferenceID;
+    }
 
-            public int getQuoteId() {
-                return QuoteId;
-            }
+    public int getQuoteId() {
+        return QuoteId;
+    }
 
-            public void setQuoteId(int QuoteId) {
-                this.QuoteId = QuoteId;
-            }
+    public void setQuoteId(int QuoteId) {
+        this.QuoteId = QuoteId;
+    }
 
-            public int getPolicyTermYear() {
-                return PolicyTermYear;
-            }
+    public int getPolicyTermYear() {
+        return PolicyTermYear;
+    }
 
-            public void setPolicyTermYear(int PolicyTermYear) {
-                this.PolicyTermYear = PolicyTermYear;
-            }
+    public void setPolicyTermYear(int PolicyTermYear) {
+        this.PolicyTermYear = PolicyTermYear;
+    }
 
-            public String getInsurerName() {
-                return InsurerName;
-            }
+    public String getInsurerName() {
+        return InsurerName;
+    }
 
-            public void setInsurerName(String InsurerName) {
-                this.InsurerName = InsurerName;
-            }
+    public void setInsurerName(String InsurerName) {
+        this.InsurerName = InsurerName;
+    }
 
-            public String getInsurerLogoName() {
-                return InsurerLogoName;
-            }
+    public String getInsurerLogoName() {
+        return InsurerLogoName;
+    }
 
-            public void setInsurerLogoName(String InsurerLogoName) {
-                this.InsurerLogoName = InsurerLogoName;
-            }
+    public void setInsurerLogoName(String InsurerLogoName) {
+        this.InsurerLogoName = InsurerLogoName;
+    }
 
-            public int getInsurerId() {
-                return InsurerId;
-            }
+    public int getInsurerId() {
+        return InsurerId;
+    }
 
-            public void setInsurerId(int InsurerId) {
-                this.InsurerId = InsurerId;
-            }
+    public void setInsurerId(int InsurerId) {
+        this.InsurerId = InsurerId;
+    }
 
-            public String getPlanName() {
-                return PlanName;
-            }
+    public String getPlanName() {
+        return PlanName;
+    }
 
-            public void setPlanName(String PlanName) {
-                this.PlanName = PlanName;
-            }
+    public void setPlanName(String PlanName) {
+        this.PlanName = PlanName;
+    }
 
-            public int getPlanID() {
-                return PlanID;
-            }
+    public int getPlanID() {
+        return PlanID;
+    }
 
-            public void setPlanID(int PlanID) {
-                this.PlanID = PlanID;
-            }
+    public void setPlanID(int PlanID) {
+        this.PlanID = PlanID;
+    }
 
-            public String getOtherPlanID() {
-                return OtherPlanID;
-            }
+    public String getOtherPlanID() {
+        return OtherPlanID;
+    }
 
-            public void setOtherPlanID(String OtherPlanID) {
-                this.OtherPlanID = OtherPlanID;
-            }
+    public void setOtherPlanID(String OtherPlanID) {
+        this.OtherPlanID = OtherPlanID;
+    }
 
-            public int getNetPremium() {
-                return NetPremium;
-            }
+    public double getNetPremium() {
+        return NetPremium;
+    }
 
-            public void setNetPremium(int NetPremium) {
-                this.NetPremium = NetPremium;
-            }
+    public void setNetPremium(double NetPremium) {
+        this.NetPremium = NetPremium;
+    }
 
-            public double getServiceTax() {
-                return ServiceTax;
-            }
+    public double getServiceTax() {
+        return ServiceTax;
+    }
 
-            public void setServiceTax(double ServiceTax) {
-                this.ServiceTax = ServiceTax;
-            }
+    public void setServiceTax(double ServiceTax) {
+        this.ServiceTax = ServiceTax;
+    }
 
-            public int getSumInsured() {
-                return SumInsured;
-            }
+    public int getSumInsured() {
+        return SumInsured;
+    }
 
-            public void setSumInsured(int SumInsured) {
-                this.SumInsured = SumInsured;
-            }
+    public void setSumInsured(int SumInsured) {
+        this.SumInsured = SumInsured;
+    }
 
-            public String getHMBValue() {
-                return HMBValue;
-            }
+    public String getHMBValue() {
+        return HMBValue;
+    }
 
-            public void setHMBValue(String HMBValue) {
-                this.HMBValue = HMBValue;
-            }
+    public void setHMBValue(String HMBValue) {
+        this.HMBValue = HMBValue;
+    }
 
-            public boolean isIsOnlinePayment() {
-                return IsOnlinePayment;
-            }
+    public boolean isIsOnlinePayment() {
+        return IsOnlinePayment;
+    }
 
-            public void setIsOnlinePayment(boolean IsOnlinePayment) {
-                this.IsOnlinePayment = IsOnlinePayment;
-            }
+    public void setIsOnlinePayment(boolean IsOnlinePayment) {
+        this.IsOnlinePayment = IsOnlinePayment;
+    }
 
-            public String getKeyFeatures() {
-                return KeyFeatures;
-            }
+    public String getKeyFeatures() {
+        return KeyFeatures;
+    }
 
-            public void setKeyFeatures(String KeyFeatures) {
-                this.KeyFeatures = KeyFeatures;
-            }
+    public void setKeyFeatures(String KeyFeatures) {
+        this.KeyFeatures = KeyFeatures;
+    }
 
-            public String getBroucherDownloadLink() {
-                return BroucherDownloadLink;
-            }
+    public String getBroucherDownloadLink() {
+        return BroucherDownloadLink;
+    }
 
-            public void setBroucherDownloadLink(String BroucherDownloadLink) {
-                this.BroucherDownloadLink = BroucherDownloadLink;
-            }
+    public void setBroucherDownloadLink(String BroucherDownloadLink) {
+        this.BroucherDownloadLink = BroucherDownloadLink;
+    }
 
-            public int getDiscount() {
-                return Discount;
-            }
+    public int getDiscount() {
+        return Discount;
+    }
 
-            public void setDiscount(int Discount) {
-                this.Discount = Discount;
-            }
+    public void setDiscount(int Discount) {
+        this.Discount = Discount;
+    }
 
-            public int getGrossPremium() {
-                return GrossPremium;
-            }
+    public int getGrossPremium() {
+        return GrossPremium;
+    }
 
-            public void setGrossPremium(int GrossPremium) {
-                this.GrossPremium = GrossPremium;
-            }
+    public void setGrossPremium(int GrossPremium) {
+        this.GrossPremium = GrossPremium;
+    }
 
-            public String getDiscountPercent() {
-                return DiscountPercent;
-            }
+    public String getDiscountPercent() {
+        return DiscountPercent;
+    }
 
-            public void setDiscountPercent(String DiscountPercent) {
-                this.DiscountPercent = DiscountPercent;
-            }
+    public void setDiscountPercent(String DiscountPercent) {
+        this.DiscountPercent = DiscountPercent;
+    }
 
-            public String getPremium() {
-                return Premium;
-            }
+    public String getPremium() {
+        return Premium;
+    }
 
-            public void setPremium(String Premium) {
-                this.Premium = Premium;
-            }
+    public void setPremium(String Premium) {
+        this.Premium = Premium;
+    }
 
-            public String getGroup_name() {
-                return Group_name;
-            }
+    public String getGroup_name() {
+        return Group_name;
+    }
 
-            public void setGroup_name(String Group_name) {
-                this.Group_name = Group_name;
-            }
+    public void setGroup_name(String Group_name) {
+        this.Group_name = Group_name;
+    }
 
-            public int getDeductible_Amount() {
-                return Deductible_Amount;
-            }
+    public int getDeductible_Amount() {
+        return Deductible_Amount;
+    }
 
-            public void setDeductible_Amount(int Deductible_Amount) {
-                this.Deductible_Amount = Deductible_Amount;
-            }
+    public void setDeductible_Amount(int Deductible_Amount) {
+        this.Deductible_Amount = Deductible_Amount;
+    }
 
-            public String getQuoteStatus() {
-                return QuoteStatus;
-            }
+    public String getQuoteStatus() {
+        return QuoteStatus;
+    }
 
-            public void setQuoteStatus(String QuoteStatus) {
-                this.QuoteStatus = QuoteStatus;
-            }
+    public void setQuoteStatus(String QuoteStatus) {
+        this.QuoteStatus = QuoteStatus;
+    }
 
-            public String getProposerPageUrl() {
-                return ProposerPageUrl;
-            }
+    public String getProposerPageUrl() {
+        return ProposerPageUrl;
+    }
 
-            public void setProposerPageUrl(String ProposerPageUrl) {
-                this.ProposerPageUrl = ProposerPageUrl;
-            }
-        }
+    public void setProposerPageUrl(String ProposerPageUrl) {
+        this.ProposerPageUrl = ProposerPageUrl;
+    }
+}
